requirements.txt
pandas
google-genai
streamlit
mysql-connector-python
faker
python-dotenv